Fixed minor issue
Also made Timer class as default in application

+++ Engine/Source/Timer.h
@@ -44,7 +44,7 @@ public:
     void SetFpsLimit(unsigned int limit) { mFpsLimit = limit; }
 
     std::vector<float> GetFpsLog() const { return mFpsLog; }
-    std::vector<unsigned long> GetMsLog() const { return mMsLog; }
+    std::vector<unsigned long long> GetMsLog() const { return mMsLog; }
 
     bool UpdateFpsLog() const { return mUpdateFpsLog; }
     void FpsLogUpdated() { mUpdateFpsLog = false; }
@@ -90,7 +90,7 @@ private:
     unsigned int mTotalFrames = 0;  //Total frames since start of Timer
 
     std::vector<float> mFpsLog;         //Log of the last 100 FPS calculated
-    std::vector<unsigned long> mMsLog;  //Log of the time of the last 100 frames
+    std::vector<unsigned long long> mMsLog;  //Log of the time of the last 100 frames
 
     bool mUpdateFpsLog = false;         //True if enough time has passed (500ms) to calculate FPS
